Decode HTML entities in hashtable index references
When `$Data.value` is sent to `ConvertTo-Html` on line 732 (line number as listed in this commit) characters that can be converted to an HTML entity are where possible.  For example:

```
å => &#229;
ä => &#228;
ö =>&#246;
' => &#39;
" => &quot;
```

In the case of DistinguishedName this causes an issue as the HTML encoded value is pulled out of a string via regex and is then used as in index to pull back the hashtable value.  Because the HTML encoded version would not match the original DistinguishedName this causes the items not to be displayed.

The above change decodes HTML entities in the DistinguishedName so that the regular expression captures/rewrites work as expected.